В методе get_vacancies_with_keyword добавлена нечувствительность к регистру
