+++ database_env.ini
@@ -1,0 +1,5 @@
+[postgresql]
+host=Имя или IP-адрес сервера
+user=Имя пользователя
+password=Пароль пользователя
+port=Порт, на котором PostgreSQL ожидает подключения
